Variable expansion into development

+++ libft/ft_getposition.c
@@ -6,13 +6,13 @@
 /*   By: emgarcia <emgarcia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/21 23:24:06 by emgarcia          #+#    #+#             */
-/*   Updated: 2021/11/22 16:55:11 by emgarcia         ###   ########.fr       */
+/*   Updated: 2021/11/23 14:40:50 by emgarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_getposition(char	*str, char c)
+size_t	ft_getposition(char	*str, char c)
 {
 	size_t	i;
 
@@ -21,6 +21,6 @@ int	ft_getposition(char	*str, char c)
 	i = -1;
 	while (str[++i])
 		if (str[i] == c)
-			return ((int)i);
+			return (i);
 	return (INT_MAX);
 }
